% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontal_plane_kinematics.R
\name{add_frontal_plane_knee_angle}
\alias{add_frontal_plane_knee_angle}
\title{Calculate frontal plane knee kinematics.
Positive values of LFPKA and RFPKA reflects lateral deviation of the knee (knee varus).}
\usage{
add_frontal_plane_knee_angle(.data)
}
\arguments{
\item{.data}{A tibble containg knee and ankle spatial joint center positions in the anatomical frontal plane. These positions can be created from
global spatial joint center positions using \code{project_full_body_to_AP()}}
}
\value{
The tibble supplied in \code{.data} argument with the added columns \code{LFPKA} and \code{RFPKA}.
}
\description{
Calculate frontal plane knee kinematics.
Positive values of LFPKA and RFPKA reflects lateral deviation of the knee (knee varus).
}
\examples{
# Prepare data
df <- data.frame(
                 LA_APR = c(10,5,0,-5,-10),
                 LK_APR = c(0,0,0,0,0),
                 LA_APU = c(0,0,0,0,0),
                 LK_APU = c(10,10,10,10,10),
                 RA_APR = c(20,15,10,5,0),
                 RK_APR = c(10,10,10,10,10),
                 RA_APU = c(0,0,0,0,0),
                 RK_APU = c(10,10,10,10,10))

add_frontal_plane_knee_angle(df)
}
