% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontal_plane_kinematics.R
\name{add_frontal_plane_projection_angle}
\alias{add_frontal_plane_projection_angle}
\title{add fontal plane projection angle kinematics to a mocap tibble}
\usage{
add_frontal_plane_projection_angle(.data)
}
\arguments{
\item{.data}{A tibble that must contain lower extremity spatial joint postions in the anatomical plane (please see \code{project_full_body_to_AP()})}
}
\value{
The tibble supplied in the .data argument with two extra columns (LFPPA and RFPPA).
}
\description{
\code{add_frontal_plane_projection_angle()} adds the frontal plane projection angle to a tibble.
\cr\cr\emph{Please note that the frontal plane projection angle has been developed for 2D video analysis of motions that do not involve the height of hip joint being
at the level of the knee joint of lower (e.g. single-leg-squats). Due to planar cross talk, the projected kinemtics produce high, and potentially misleading, values when the height of the hip joint
approaches that of the knee joint. The frontal plane projeciton kinematics are likely only usefull for analyzing motions that do
not involve deep positions of the pelvis (e.g. single-leg squats, gait etc.). \cr Please see examples for further detail.}
}
\examples{
# Prepare data
df <- data.frame(
                 LH_APR = c(10,10,10, 0,10,20),
                 LK_APR = c(10,10,10,10,10,10),
                 LA_APR = c( 0,10,20,10,10,10),
                 LH_APU = c(30,30,30,30,30,30),
                 LK_APU = c(10,10,10,10,10,10),
                 LA_APU = c( 0, 0, 0, 0, 0, 0),
                 RH_APR = c(10,10,10, 0,10,20),
                 RK_APR = c(10,10,10,10,10,10),
                 RA_APR = c( 0,10,20,10,10,10),
                 RH_APU = c(30,30,30,30,30,30),
                 RK_APU = c(10,10,10,10,10,10),
                 RA_APU = c( 0, 0, 0, 0, 0, 0))

add_frontal_plane_projection_angle(df)

# Example to view L and R FPPA
df <- mocapr::mocapr_synthetic_data
df <- dplyr::filter(df, sample == 2)
df <- add_frontal_plane_projection_angle(df)
animate_anatomical(df,
                  planes = "R",
                  return_plot = TRUE,
                  col_facets = col_fa,
                  row_facets = row_fa)+
   ggplot2::geom_text(ggplot2::aes(label = paste0("LFPPA: ", round(LFPPA, 1)),
                      x = mean(value)-15, y= max(U)+10), color = "black", size = 3)+
   ggplot2::geom_text(ggplot2::aes(label = paste0("RFPPA: ", round(RFPPA, 1)),
                      x = mean(value)+15, y= max(U)+10), color = "black", size = 3)

# Explore the planar cross talk at deep positions of the hip-joint centers.
df2 <- dplyr::filter(mocapr::mocapr_synthetic_data, sample == 2)
df2 <- add_frontal_plane_knee_angle(df2)
df2 <- add_frontal_plane_knee_deviation(df2)
df2 <- add_frontal_plane_projection_angle(df2)
animate_anatomical(df2,
                  planes = "R",
                  return_plot = TRUE,
                  col_facets = col_fa,
                  row_facets = row_fa)+
ggplot2::geom_text(ggplot2::aes(label = paste0("LFPPA: ", round(LFPPA, 1)),
                  x = mean(value)-15, y= max(U)+30), color = "black", size = 3)+
ggplot2::geom_text(ggplot2::aes(label = paste0("RFPPA: ", round(RFPPA, 1)),
                  x = mean(value)+15, y= max(U)+30), color = "black", size = 3)+
ggplot2::geom_text(ggplot2::aes(label = paste0("LFPKA: ", round(LFPKA, 1)),
                  x = mean(value)-15, y= max(U)+20), color = "black", size = 3)+
ggplot2::geom_text(ggplot2::aes(label = paste0("RFPKA: ", round(RFPKA, 1)),
                  x = mean(value)+15, y= max(U)+20), color = "black", size = 3)+
ggplot2::geom_text(ggplot2::aes(label = paste0("LFPKD: ", round(LFPKD, 1)),
                  x = mean(value)-15, y= max(U)+10), color = "black", size = 3)+
ggplot2::geom_text(ggplot2::aes(label = paste0("RFPKD: ", round(RFPKD, 1)),
                  x = mean(value)+15, y= max(U)+10), color = "black", size = 3)

}
