% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_functions.R
\name{add_jump_events}
\alias{add_jump_events}
\title{add_jump_events()}
\usage{
add_jump_events(.data)
}
\arguments{
\item{.data}{A tibble containing motioncapture data from a squat in the format used in the mocapr package.}
}
\value{
A tibble
}
\description{
add_jump_events()
}
\examples{
\dontrun{}
}
