% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontal_plane_kinematics.R
\name{add_knee_anke_hip_ratios}
\alias{add_knee_anke_hip_ratios}
\title{Add knee-to-hip, ankle-to-hip, and knee-to-ankle separation distance ratios.}
\usage{
add_knee_anke_hip_ratios(.data)
}
\arguments{
\item{.data}{A tbbile containg ankle, knee, and hip spatial joint center positions in the anatomical frontal plane. These positions can
be generated from global spatial positions using \code{project_full_body_to_AP()}.}
}
\value{
The tibble supplied in \code{.data} argument with the added columns \code{KHR}, \code{AHR}, and \code{KAR}.
}
\description{
The ratios are calculated in the anatomical frontal plane. The ratios are calculated using both lower extremities. Therefore, they
should only be used to analyze bilateral symetric movements (e.g., squats, or bilateral jump-landings).
}
\examples{
# Prepare data
df <- dplyr::filter(mocapr::mocapr_synthetic_data, sample == 1)
add_knee_anke_hip_ratios(df)

# The ratio measures are unaffected by the depth of the pelvis.

df2 <- dplyr::filter(mocapr::mocapr_synthetic_data, sample == 2)
df2 <- add_knee_anke_hip_ratios(df2)
animate_anatomical(df2,
                  planes = "R",
                  return_plot = TRUE,
                  col_facets = col_fa,
                  row_facets = row_fa)+
ggplot2::geom_text(ggplot2::aes(label = paste0("KHR: ", round(KHR, 2)),
                  x = mean(value), y= max(U)+30), color = "black", size = 3)+
ggplot2::geom_text(ggplot2::aes(label = paste0("AHR: ", round(AHR, 2)),
                  x = mean(value), y= max(U)+20), color = "black", size = 3)+
ggplot2::geom_text(ggplot2::aes(label = paste0("KAR: ", round(KAR, 2)),
                  x = mean(value), y= max(U)+10), color = "black", size = 3)
}
