% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_functions.R
\name{add_phases_jump}
\alias{add_phases_jump}
\title{Divide a jump into phases}
\usage{
add_phases_jump(.data)
}
\arguments{
\item{.data}{A tibble containing motioncapture data from a jump. The data must contain:\cr
\itemize{
\item \code{LHY} and \code{RHY} columns containg global spatial joint-center positions of the left and right hip\cr
\item A \code{frame} column\cr
\item A \code{marks} column with values descring toe off events (TOL = toe off of left foot, TOR = toe off of right foot, TOB = toe off from both feet),
and impact events (FFL = flat foot contact with left foot, FFR = flat foot contact with right foot, FFB = flat foot contact with both feet).
}}
}
\value{
The tibble supplied in the \code{.data argument}, with the added columns \code{phase_l}, \code{phase_r}, \code{phase_b},
\code{events_l}, \code{events_r}, \code{events_b}
}
\description{
\code{add_phases_jump} uses average hip joint-center height and supplied events to divide a jump into 11 phases.
This division is useful for further analyis of the individual phases in the jump.
}
\details{
Three phase columns are generated \code{phase_l, phase_r, phase_b}, describing the phases of the left leg, right leg, and both extremities respectively.

The numerical values in \code{phase} column correspond to:
\enumerate{
\item Descending phase of the preparation
\item The deepest postion in the preparation phase (A single frame)
\item The ascending phase of the preparation (push-off)
\item Toe off - the last fase with contact between the feet and the ground (A single frame)
\item The ascending phase of the hang time
\item The highest position of the hang time (A single frame)
\item The descending phase of the hang time
\item Impact (A single frame)
\item Descending phase of the landing.
\item Deepest position of the landing (A single frame)
\item Ascending phase of the landing.
}

The values in the column \code{phase_b} will be identical to the values in \code{phase_l} and \code{phase_r} if the subject is taking off
and landing simultaneously with both extremities. If this is not the case \code{phase_b} will take the value \code{4} at the frame where
the last foot has contact with the ground, and the value \code{8} at the frame where the last foot has impact. \code{NA} values will be given to phases
where there is transition (e.g. left foot still has contact with ground, while right foot is in the air.)
}
\examples{
# With synthetic data
df <- tibble::tibble(
       frame = c(1:11),
       marks = c( NA, NA, NA,"TOB",  NA,  NA,  NA, "FFB", NA,  NA,  NA),
       LHY =   c(100, 50, 75,  100, 125, 150, 125,   100, 75,  50, 100),
       RHY =   c(100, 50, 75,  100, 125, 150, 125,   100, 75,  50, 100))

df2 <- tibble::tibble(
        frame = c(1:15),
        marks = c( NA, NA, NA, "TOL", "TOR",  NA,  NA,  NA,  NA, "FFL", "FFR",  NA,  NA, NA,  NA),
        LHY =   c(100, 50, 50,    75,  100,  125, 150, 150, 125,   100,    75,  70,  50, 50, 100),
        RHY =   c(100, 50, 50,    75,  100,  125, 150, 150, 125,   100,    75,  70,  50, 50, 100))
add_phases_jump(df)
add_phases_jump(df2)

# With real mocap data
 df3 <- dplyr::filter(mocapr::mocapr_data, movement_nr == 1)
 df3 <- dplyr::select(df, frame, marks, LHY, RHY)

 add_phases_jump(df3)

# A plot displaying the phases
df4 <- dplyr::filter(mocapr::mocapr_data, movement_nr == 1)
df4 <- add_phases_jump(df4)
 mocapr::animate_global(df4,
                        planes = "X",
                        remove_facet_labels = FALSE,
                        return_plot = TRUE,
                        col_facets = phase_b)
}
