% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squat_functions.R
\name{add_squat_events}
\alias{add_squat_events}
\title{Divide a squat movement into phases and events}
\usage{
add_squat_events(.data)
}
\arguments{
\item{.data}{A tibble containing motioncapture data from a squat in the format used in the mocapr package.}
}
\value{
A tibble with three added columns. "squat_events", "phase", and "squat_depth"
}
\description{
\code{add_squat_events()} takes a tibble as the only argument, and returns the tible with three extra columns usefull for further analysis.
The three extra columns are: "squat_events", "phase", and "squat_depth".
The tibble must contain motioncapture data in format used in mocapr package,
and the captured motion must be a squat. Please note that the function will not fail if the movement is not a squat.
}
\examples{
# Prepare data
df <- dplyr::filter(mocapr::mocapr_data, movement_nr == 1, frame > 20, frame < 70)

add_squat_events(df)
}
