% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Functions.R
\name{animate_global}
\alias{animate_global}
\title{animate_global().
Please see GitHub README.me for a more detailed description.}
\usage{
animate_global(.data, animate = TRUE, ...)
}
\arguments{
\item{.data}{A tibble containing global joint center positions (X, Y, Z).}

\item{animate}{Defaults to TRUE. If false the function will provide a plot faceted on frames.}

\item{...}{These parameters are passed to the gganimate::animate() function}
}
\value{
Defaults to an animated gif. Different outputs can be achieved by passing different arguments via ... to the gganimate::animate() function. If animate = FALSE a ggplot plot is returned.
}
\description{
animate_global().
Please see GitHub README.me for a more detailed description.
}
\examples{
mocapr::animate_global(dplyr::filter(mocapr::mocapr_data, movement_nr == 1), fps = 50)
\dontrun{
jump_1 <-  dplyr::filter(mocapr::mocapr_data, movement_nr == 1)
mocapr::animate_global(jump_1, nframes = nrow(jump_1), fps = 50)
}
}
