% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_captury_csv}
\alias{import_captury_csv}
\title{Import csv files exported by the Captury Live system version 0.0.168 or later}
\usage{
import_captury_csv(filename, frames_pr_second = 50)
}
\arguments{
\item{filename}{Path and filename of a .csv file containing motion capture data from the Captury system}

\item{frames_pr_second}{Recorded frames pr. second used in the setup when capturing the data. Defaults to 50.}
}
\value{
A tibble containing joint angles and global joint center positions of the: toes, ankles, knees, hips, center of gravity, shoulders, elbows, and wrists.
}
\description{
\code{import_captury_csv} takes the filepath and filename of a .csv file containg motion capture data captured and exported using the CapturyLive motion capture system. The
.csv file is then imported and cleaned and returned as a tibble. All joint angles and global joint center positions are in abbreviated names (e.g. left knee flexion =
LKF, global Y coordinate of the right hip joint is RHY).\cr
Please see the GitHub README.me for a more detailed description.
}
\examples{
path <- system.file("examples", "SBJ_new_version.csv", package = "mocapr")
import_captury_csv(path)
}
