% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_DARI_bvh}
\alias{import_DARI_bvh}
\title{import_DARI_bvh}
\usage{
import_DARI_bvh(filename)
}
\arguments{
\item{filename}{Path and file-name of the .bvh file to be imported}
}
\value{
A tibble containing global joint center positions of the: toes, ankles, knees, hips, clavicles, shoulders, elbows, and wrists.
}
\description{
import_DARI_bvh takes the file-path and file-name of a .bvh file containing motion capture data exported using the DARI motion capture system. The
.bvh file is then imported, cleaned, and returned as a tibble. All global joint center positions are in abbreviated names (e.g. global Y coordinate of the right hip joint is RHY).\cr
Please see the GitHub README.me for a more detailed description.
}
\section{Note}{

This function imports .bvh files as they are exported from the DARI motion capture software.
}

\examples{
\dontrun{
path <- system.file("examples", "DARI.bvh", package = "mocapr")
suppressMessages(import_DARI_bvh(path))
}
}
