% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_freemocap_csv}
\alias{import_freemocap_csv}
\title{FreeMoCap import
\code{import_freemocap_csv} takes the filepath and filename of a .csv file containing motion capture data captured and exported using the FreeMoCap motion capture system. The
.csv file is then imported and cleaned and returned as a tibble. All global joint center positions are in abbreviated names (e.g. global Y coordinate of the right hip joint is RHY).\cr
Please see the GitHub README.me for a more detailed description.}
\usage{
import_freemocap_csv(
  filename,
  keep_face_coords = FALSE,
  keep_finger_coords = FALSE
)
}
\arguments{
\item{filename}{Path and filename to a .csv file with MoCap data from the FreeMoCap software}

\item{keep_face_coords}{A logical(TRUE/FALSE). Should coordinates from ears, mouth, eyes, and nose be kept if they exist in the dataframe? Defaults to FALSE}

\item{keep_finger_coords}{A logical(TRUE/FALSE). Should coordinates from fingers be kept if they exist in the dataframe? Defaults to FALSE}
}
\value{
A tibble
}
\description{
FreeMoCap import
\code{import_freemocap_csv} takes the filepath and filename of a .csv file containing motion capture data captured and exported using the FreeMoCap motion capture system. The
.csv file is then imported and cleaned and returned as a tibble. All global joint center positions are in abbreviated names (e.g. global Y coordinate of the right hip joint is RHY).\cr
Please see the GitHub README.me for a more detailed description.
}
\examples{
path <- system.file("examples", "freemocap_sit_to_stand.csv", package = "mocapr")
import_freemocap_csv(path)
}
