% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_optitrack_csv}
\alias{import_optitrack_csv}
\title{Import csv files exported by the Optitrack system}
\usage{
import_optitrack_csv(filename)
}
\arguments{
\item{filename}{Path and filename of a .csv file containg motion capture data from the Optitrack system}
}
\value{
A tibble containing joint angles and global joint center positions of the: toes, ankles, knees, hips, center of gravity, shoulders, elbows, and wrists.
}
\description{
\code{import_optitrack_csv} takes the filepath and filename of a .csv file containg motion capture data captured and exported using the Optitrack motion capture system. The
.csv file is then imported and cleaned and returned as a tibble. All joint angles and global joint center positions are in abreviated names (e.g. left knee flexion =
LKF, global Y coordinate of the right hip joint is RHY).\cr
Please see the GitHub README.me for a more detailed description.
}
\examples{
path <- system.file("examples", "optitrack.csv", package = "mocapr")
import_optitrack_csv(path)
}
