% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_functions.R
\name{mocap_finish_animation}
\alias{mocap_finish_animation}
\title{convert a mocap plot in the mocapr format to an animation}
\usage{
mocap_finish_animation(.data, ...)
}
\arguments{
\item{.data}{A ggplot2 object containing mocap data in the mocapr format.}

\item{...}{Additional arguments are passed to \code{\link[gganimate]{animate}}.}
}
\value{
A gif
}
\description{
convert a mocap plot in the mocapr format to an animation
}
\examples{
df <- dplyr::filter(mocapr::mocapr_data, movement_nr == 1)
df <- mocapr::project_full_body_to_AP(df)
df <- mocapr::animate_anatomical(df,
                                 return_data = TRUE)
df <- mocapr::mocap_plot_basic(df)
df <- mocapr::mocap_plot_avatar(df)
# mocap_finish_animation(df)
}
