% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Functions.R
\docType{data}
\name{mocapr_data}
\alias{mocapr_data}
\alias{import_captury}
\title{This is data to be included in my package}
\usage{
import_captury(filename, frames_pr_second = 50)
}
\arguments{
\item{filename}{Path and filename of a .csv file containg motion capture data from the Captury system}

\item{frames_pr_second}{Recorded frames pr. second used in the setup when capuring the data. Defaults to 50.}
}
\value{
A tibble containg joint angles and global joint center positions of the: toes, ankles, knees, hips, center of gravity, shoulders, elbows, and wrists.
}
\description{
This is data to be included in my package
}
\examples{
\dontrun{}
}
\references{
\url{https://www.youtube.com/playlist?list=PLMjrjny4Ymmd1nSGHU0A6dWfEWjBxc-VQ}
}
\author{
Steen Harsted \email{steenharsted@gmail.com}
}
\keyword{(e.g.}
\keyword{.csv}
\keyword{=}
\keyword{All}
\keyword{CapturyLive}
\keyword{GitHub}
\keyword{Import_Captury()}
\keyword{LKF,}
\keyword{README.me}
\keyword{RHY).Please}
\keyword{The}
\keyword{Y}
\keyword{a}
\keyword{abreviated}
\keyword{and}
\keyword{angles}
\keyword{are}
\keyword{as}
\keyword{capture}
\keyword{captured}
\keyword{center}
\keyword{cleaned}
\keyword{containg}
\keyword{coordinate}
\keyword{data}
\keyword{description.}
\keyword{detailed}
\keyword{exported}
\keyword{file}
\keyword{filename}
\keyword{filepath}
\keyword{flexion}
\keyword{for}
\keyword{global}
\keyword{hip}
\keyword{import_captury}
\keyword{imported}
\keyword{in}
\keyword{is}
\keyword{joint}
\keyword{knee}
\keyword{left}
\keyword{more}
\keyword{motion}
\keyword{names}
\keyword{of}
\keyword{positions}
\keyword{returned}
\keyword{right}
\keyword{see}
\keyword{system.}
\keyword{takes}
\keyword{the}
\keyword{then}
\keyword{tibble.}
\keyword{using}
