% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_anatomical}
\alias{plot_anatomical}
\title{animate_anatomical()}
\usage{
plot_anatomical(.data, animate = TRUE, planes = c("R", "F"),
  use_geom_point = TRUE, extra_row_facets = NULL, ...)
}
\arguments{
\item{.data}{A tibble containing joint center positions in the anatomical planes generated using the project_to_AP() function.}

\item{animate}{Defaults to TRUE. If false the function will provide a plot faceted on frames.}

\item{use_geom_point}{Defaults to TRUE. If TRUE, points in the animation will be created using ggplot2::geom_point(). If FALSE, points are created using ggforce::geom_circle(). This will ensure correct size-proportions of the points, and that no points are croped out of the animation. However, it comes at the prize of much longer rendering-times.}

\item{extra_row_facets}{A test}

\item{...}{These arguments are passed to the gganimate::animate() function.}
}
\value{
Defaults to an animated gif. Different outputs can be achieved by passing different arguments via ... to the gganimate::animate() function. If animate = FALSE a ggplot plot is returned.
}
\description{
animate_anatomical()
}
\examples{
\dontrun{}
}
