% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_movement}
\alias{plot_movement}
\title{animate_movement().
Please see GitHub README.me for a more detailed description.}
\usage{
plot_movement(.data, animate = TRUE, use_geom_point = TRUE, ...)
}
\arguments{
\item{.data}{A tibble containing joint center positions in the movement plane generated using the project_to_MP() function.}

\item{animate}{Defaults to TRUE. If false the function will provide a plot faceted on frames.}

\item{use_geom_point}{Defaults to TRUE. If TRUE, points in the animation will be created using ggplot2::geom_point(). If FALSE, points are created using ggforce::geom_circle(). This will ensure correct size-proportions of the points, and that no points are croped out of the animation. However, it comes at the prize of much longer rendering-times.}

\item{...}{These parameters are passed to the gganimate::animate() function}
}
\value{
Defaults to an animated gif. Different outputs can be achieved by passing different arguments via ... to the gganimate::animate() function. If animate = FALSE a ggplot plot is returned.
}
\description{
animate_movement().
Please see GitHub README.me for a more detailed description.
}
\examples{
\dontrun{}
}
