% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Functions.R
\name{project_full_body_to_AP}
\alias{project_full_body_to_AP}
\title{project_full_body_to_AP()
project_full_body_to_AP() uses project_single_joint_to_AP() to project a pre-specified collection of joint centers onto the anatomical planes of the subject.
The pre-specified joint centers are the following from the left and right side: toe, ankle, knee, hip, wrist, elbow, and shoulder.
Please see the GitHub README.md fpr a more detailed description.}
\usage{
project_full_body_to_AP(.data)
}
\arguments{
\item{.data}{A tibble containing 3D positions of the following left and right joints: toe, ankle, knee, hip, wrist, elbow, and shoulder}
}
\value{
A tibble with the positions of the pre-specied joint-centers in the anatomical planes.
}
\description{
project_full_body_to_AP()
project_full_body_to_AP() uses project_single_joint_to_AP() to project a pre-specified collection of joint centers onto the anatomical planes of the subject.
The pre-specified joint centers are the following from the left and right side: toe, ankle, knee, hip, wrist, elbow, and shoulder.
Please see the GitHub README.md fpr a more detailed description.
}
\examples{
dontrun{}
}
