% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Functions.R
\name{project_single_joint_to_AP}
\alias{project_single_joint_to_AP}
\title{project_single_joint_to_AP()}
\usage{
project_single_joint_to_AP(.data, Y, X, Z, New_Name = "New")
}
\arguments{
\item{.data}{A tibble containing the global 3D positions of the joints given in the parameters X, Y, Z and the 3D  positions of both hip joints.}

\item{Y}{The name of the global Y coordinate column (up direction) of the joint you wish to project to the frontal plane}

\item{X}{The name of the global X coordinate column of the joint you wish to project to the frontal plane}

\item{Z}{The name of the global Z coordinate column of the joint you wish to project to the frontal plane}

\item{New_Name}{The abreviated name of the new joint, the name of the returned variables will start with the value given in New_Name}
}
\value{
A tibble containing two columns with coordinates in the right and up direction. The variables are named '"New_Name"_FPR' and '"New_Name"_FPU'
}
\description{
project_single_joint_to_AP() takes the global 3D coordinates of a single joint and project these coordinates onto the anatomical planes of the subject (the frontal and the sagital plane).
The frontal plane is defined as the plane between the two hip joint centers that is perpendicular to the floor. The sagital plane is perpendicular to the frontal
and floor plane.\cr
As the subject moves the anatomical planes will change with each frame as the pose of the subject changes. This is different to the movement planes created
by the project_single_joint_to_MP() function where the planes stay the same througout the movement.
Please see the GitHub README.me for a more detailed description.
}
\examples{
\dontrun{}
}
