% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Functions.R
\name{Project_to_MP}
\alias{Project_to_MP}
\title{Project_to_MP}
\usage{
Project_to_MP(.data, Y, X, Z, New_Name = "New")
}
\arguments{
\item{.data}{A tibble containing the global 3D positions of the joints given in the parameters X, Y, Z and the 3D  positions of both hip joints.}

\item{Y}{The name of the global Y coordinate column (up direction) of the joint you wish to project to the movement plane}

\item{X}{The name of the global X coordinate column of the joint you wish to project to the movement plane}

\item{Z}{The name of the global Z coordinate column of the joint you wish to project to the movement plane}

\item{New_Name}{The abreviated name of the new joint, the name of the returned variables will start with the value given in New_Name}
}
\value{
A tibble containig three columns with coordinates in the forward, up, and right direction. The variables are named '"New_Name"_MPF', '"New_Name"_MPU' and '"New_Name"_MPR'
}
\description{
Project_to_MP() projects the global joint center positions onto the movement plane (MP). MP is calculated by first creating a direction going from
the position of the hip joint centers at the first frame to the position of the hip joint centers at the last frame. Please see the vignette for a
more in-depth explanation.
}
\examples{
dontrun{}
}
