% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_functions.R
\name{add_jump_length_and_height}
\alias{add_jump_length_and_height}
\title{Calculate measures of jump length and jump height}
\usage{
add_jump_length_and_height(.data, method = "movement_plane")
}
\arguments{
\item{.data}{A tibble containing motion-capture data from a jump.\cr
The data must contain the following columns:
\itemize{
\item \code{jump_events} You can create this column using \code{\link{add_jump_events}}
\item \code{marks} A character column containg one or more of c("TOL", "TOR", "TOB"), AND one or more of c("FFL", "FFR", "FFB")
\item Global spatial ankle joint center positions in the floor plane: \code{LAX} \code{LAZ} \code{RAX} \code{RAZ}
\item Global spatial hip joint-center height positions: \code{LHY} \code{RHY}
}}

\item{method}{The method that should be used to calculate jump length.}
}
\value{
The tibble suplied in \code{.data} argument with the added columns \code{jump_length} and \code{jump_height} both measures are in cm.
}
\description{
Adds the columns \code{jump_length} and \code{jump_height} to a tibble containg mocap data of a jump.
}
\examples{
# Prepare data
df <- dplyr::filter(mocapr::mocapr_data, movement_nr == 1)
df <- add_phases_jump(df)
df <- mocapr::project_full_body_to_MP(df)

add_jump_length_and_height(df)

}
