% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_functions.R
\name{animate_anatomical}
\alias{animate_anatomical}
\title{animate_anatomical() animates motioncapture data that is projected onto the anatomical planes of the body using mocapr::project_full_body_to_AP().}
\usage{
animate_anatomical(.data, planes = c("R", "F"),
  planes_in_rows_or_cols = c("cols"), use_geom_point = TRUE,
  row_facets = NULL, col_facets = NULL, remove_facet_labels = TRUE,
  line_size = 2, point_size = 4, circle_size = 40,
  head_scale = 1.5, torso_scale = 1.5, return_data = FALSE,
  return_plot = FALSE, preserve = NULL, ...)
}
\arguments{
\item{.data}{A tibble containing joint center positions in the anatomical planes generated using the project_to_AP() function.}

\item{planes}{Planes to animate. Must be one or both of c("R", "F"). The planes are created by combining the suplied direction with the up-direction. Defaults to c("R", "F").}

\item{planes_in_rows_or_cols}{Facet the chosen planes in either rows or columns. Must be one of c("rows", "cols"). Defaults to "cols".}

\item{use_geom_point}{Defaults to TRUE. If TRUE, points in the animation will be created using ggplot2::geom_point(). If FALSE, points are created using ggforce::geom_circle(). This will ensure correct size-proportions of the points, and that no points are croped out of the animation. However, it comes at the prize of much longer rendering-times.}

\item{row_facets}{Make additional row-facets in the animation using a given variable. Defaults to NULL.}

\item{col_facets}{Make additional column-facets in the animation using a given variable. Defaults to NULL.}

\item{remove_facet_labels}{Remove the facet labels. Defaults to TRUE.}

\item{line_size}{The size of the lines connecting the joint centers.}

\item{point_size}{The size of the joint centers when use_geom_point = TRUE.}

\item{circle_size}{The size of the joint centers when use_geom_point = FALSE.}

\item{head_scale}{The size of the head relative to the joint-centers.}

\item{torso_scale}{The size of the torso line relative to the remaing lines.}

\item{return_data}{Return the wrangled data before it is sent to ggplot. This is useful for understanding the data-structure and for developmental purposes.}

\item{return_plot}{Return a plot instead of an animaiton. This is useful for customizing the plot before passing it to gganimate::}

\item{preserve}{Which variables should be preserved. This is usefull if you wish to add additional layers to the animations e.g., geom_text()}

\item{...}{These arguments are passed to the gganimate::animate() function.}
}
\value{
Defaults to an animated gif. Different outputs can be achieved by passing different arguments via ... to the gganimate::animate() function. If return_plot = TRUE a ggplot plot is returned.
}
\description{
animate_anatomical() animates motioncapture data that is projected onto the anatomical planes of the body using mocapr::project_full_body_to_AP().
}
\examples{
\dontrun{}
}
