% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Functions.R
\name{animate_anatomical}
\alias{animate_anatomical}
\title{animate_anatomical()}
\usage{
animate_anatomical(.data, ...)
}
\arguments{
\item{.data}{A tibble containing joint center positions in the anatomical planes generated using the project_to_AP() function.}

\item{...}{These arguments are passed to the gganimate::animate() function.}
}
\value{
An animated gif
}
\description{
animate_anatomical()
}
\examples{
dontrun{}
}
