% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Functions.R
\name{animate_movement}
\alias{animate_movement}
\title{animate_movement()}
\usage{
animate_movement(.data, ...)
}
\arguments{
\item{.data}{A tibble containing joint center positions in the movement plane generated using the project_to_MP() function.}

\item{...}{These parameters are passed to the gganimate::animate() function}
}
\value{
An animated gif
}
\description{
animate_movement()
}
\examples{
dontrun{}
}
