% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_captury}
\alias{import_captury}
\title{import_captury}
\usage{
import_captury(filename, frames_pr_second = 50)
}
\arguments{
\item{filename}{Path and file name of a .csv file containing motion capture data from the Captury system}

\item{frames_pr_second}{Recorded frames pr. second used in the setup when capturing the data. Defaults to 50.}
}
\value{
A tibble containing joint angles and global joint center positions of the: toes, ankles, knees, hips, center of gravity, shoulders, elbows, and wrists.
}
\description{
import_captury() takes the file path and file name of a .csv file containing motion capture data captured and exported using the CapturyLive motion capture system. The
.csv file is then imported and cleaned and returned as a tibble. All joint angles and global joint center positions are in abbreviated names (e.g. left knee flexion =
LKF, global Y coordinate of the right hip joint is RHY).\cr
Please see the GitHub README.me for a more detailed description.
}
\section{Note}{

This function imports Captury .csv files as they are exported from CapturyLive version 0.0.168 and earlier. Newer versions have a different layout
and should be imported using \code{import_captury_csv()}.
}

\examples{
path <- system.file("examples", "vertical_jump_old_cap_version.csv", package = "mocapr")
suppressMessages(import_captury(path))
}
