% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Functions.R
\name{Import_Captury}
\alias{Import_Captury}
\title{Import_Captury}
\usage{
Import_Captury(filename)
}
\arguments{
\item{filename}{Path and filename of a .csv file containg motion capture data from the Captury system}
}
\value{
A tibble containg joint angles and global joint center positions of the: toes, ankles, knees, hips, center of gravity, shoulders, elbows, and wrists.
}
\description{
Import_Captury() takes the filepath and filename of a .csv file containg motion capture data captured and exported using the CapturyLive motion capture system. The
.csv file is then imported and cleaned and returned as a tibble. All joint angles and global joint center positions are in abreviated names (e.g. left knee flexion =
LKF, global Y coordinate of the right hip joint is RHY).Please see the package vignette for a more detailed description.
}
\examples{
dontrun{}
}
