% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_functions.R
\name{project_full_body_to_MP}
\alias{project_full_body_to_MP}
\title{project_full_body_to_MP()
project_full_body_to_MP() uses project_single_joint_to_MP() to project a pre-specified collection of joint centers onto the movement planes of the subject.
The pre-specified joint centers are the following from the left and right side: toe, ankle, knee, hip, wrist, elbow, and shoulder.
Please see the GitHub README.md fpr a more detailed description.}
\usage{
project_full_body_to_MP(.data)
}
\arguments{
\item{.data}{A tibble containing 3D positions of the following left and right joints: toe, ankle, knee, hip, wrist, elbow, and shoulder}
}
\value{
A tibble with the positions of the pre-specied joint-centers in the movement planes.
}
\description{
project_full_body_to_MP()
project_full_body_to_MP() uses project_single_joint_to_MP() to project a pre-specified collection of joint centers onto the movement planes of the subject.
The pre-specified joint centers are the following from the left and right side: toe, ankle, knee, hip, wrist, elbow, and shoulder.
Please see the GitHub README.md fpr a more detailed description.
}
\examples{
df <- dplyr::filter(mocapr::mocapr_data, movement_nr == 1)
project_full_body_to_MP(df)
}
